screen: improve message reference resolving
The current implementation has issues when yet uncached thread messages
are updated. In this case two [???] placeholders are present in the
message and it is undefined which one will be filled in first, the
thread reference or the update reference. Actually it turned out to be
the wrong order, therefore resulting in wrong information.

To fix this fully reformat messages waiting for message references.
Rename the involved functions and data structures from "thread" to
"message references", since this functionality has become more generic
by now. The original message that contained unresolved references also
needs to be stored in the waiters map now, because for incremental
updates we need the original message state we saw first.

This also ends the need to fiddle in Urwids internal object state to fix
up the placeholders.