feature: recognize lost API connection and display error message
Sometimes when there is no protocol activity over a longer time period,
the realtime API connection is lost. The reasons are unclear, because a
PING protocol is running on the web socket that should prevent this.

To at least notify the user about this condition display a red error
message that informs about this. In this state no new messages will be
received, any attempts to execute operations will fail with "connection
already closed".