bugfix: fix KeyError when a direct message is received for the first time
When another user sends a message to us that never talked to us before
then we're not yet subscribed to the respective direct chat. In this
situation a KeyError occurs when trying to get the current room state.

Therefore check if we have a room state yet and if not, simply do
nothing.