screen: display feedback when processing large room history cache
Currently when selecting a room for which a large history has been
previously cached then UI will freeze for a longer time without
feedback. Therefore show a status message about what is going on.

To prevent these long processing times, the rendered messages have to be
cached in the future, but this requires larger design changes.