feature: treat rooms with unread messages as ATTENTION state
This allows to recognize new messages in direct chats or messages that
mention the logged in user, when they are received while the client is
offline.

This relies on the server side to correctly keep track of the unread
count. There seems to be a tendency that the server can't correctly
reset the unread state in some circumstances. Especially the unread
thread handling seems broken. Ignore that part for the moment for this
reason.